chore: Renamed ExecutionContextService to ExecutionContext.
