+++ packages/architectura/test/Core/Base/BasePreHook.spec.mts
@@ -1,6 +1,6 @@
 import { describe, it } from "node:test";
 
-import { BasePreHook } from "../../../src/index.mjs";
+import { BasePreHook } from "../../../src/_index.mjs";
 
 describe(
     "BasePreHook",
