+++ packages/architectura/test/Core/Executor.spec.mts
@@ -1,6 +1,6 @@
 import { describe, it } from "node:test";
 
-import { Executor } from "../../src/index.mjs";
+import { Executor } from "../../src/_index.mjs";
 
 describe(
     "Executor",
