+++ packages/architectura/test/Core/Time.spec.mts
@@ -3,7 +3,7 @@ import { default as assert } from "node:assert/strict";
 
 import { describe, it } from "node:test";
 
-import { Time, TimeFormattingEnum } from "../../src/index.mjs";
+import { Time, TimeFormattingEnum } from "../../src/_index.mjs";
 
 describe(
 	"Time",
