+++ packages/architectura/test/Hooks/CookiePreHook.spec.mts
@@ -1,6 +1,6 @@
 import { describe, it } from "node:test";
 
-import { CookiePreHook } from "../../src/index.mjs";
+import { CookiePreHook } from "../../src/_index.mjs";
 
 describe(
     "CookiePreHook",
